+++ src/index.ts
@@ -1,0 +1,2 @@
+import MyPopupModal from "./PopupModal";
+export default MyPopupModal;
